Move stats to own section in config
