+++ .gitignore
@@ -36,6 +36,7 @@ ops.json
 help.yml
 server.properties
 sportpaper.yml
+sportpaper.jar
 update
 usercache.json
 whitelist.json
